CoM param set by dash and pilot_input is an increment
Previously the CoM offset parameter was the absolute CoM of the robot to be used by thrust. The issue I was running into is thrust has the param center_of_mass_offset which is modified by both dash andpilot_input. Two nodes modifying the same parameter is a weird issue because in order to do this you have to see if the value has changed from the previously known value to that node, then modify the value and set the param. Making the param the increment requested by these nodes eliminates that problem