CoM with two parameters this
This commit contains code to shift the CoM from multiple nodes with increments. This solution uses two parameters, it would be nice to do it with one. Especially because this does not work. When a node sets its own parameter, it does so silently, meaning it is not reported over /parameter_events. This breaks the reasoning behind the two parameter implementation. Need to find a working solution